+++ src/source/ofxAEVideoSource.h
@@ -13,7 +13,7 @@ class VideoSource : public LayerSource
 public:
 	void accept(Visitor &visitor) override;
 	bool load(const std::filesystem::path &filepath) override;
-	bool setFrame(int frame) override;
+	bool setFrame(float frame) override;
 	void draw(float x, float y, float w, float h) const override;
 	float getWidth() const override;
 	float getHeight() const override;
